@@ -23,6 +23,7 @@ from tensorflow.python.util import deprecation
 tf.compat.v1.logging.set_verbosity(tf.compat.v1.logging.ERROR)
 tf.get_logger().setLevel(logging.ERROR)
 
+
 # Monkey patching deprecation utils to supress warnings.
 # noinspection PyUnusedLocal
 def deprecated(
@@ -49,7 +50,7 @@ from tensorflow.keras.regularizers import l1_l2
 # Custom Modules
 try:
     from ..neural_network_methods import NeuralNetworkMethods
-except (ModuleNotFoundError, ValueError):
+except (ModuleNotFoundError, ValueError, ImportError):
     from impute.unsupervised.neural_network_methods import NeuralNetworkMethods
 
 
@@ -237,7 +238,6 @@ class UBPPhase1(tf.keras.Model):
         ToDo:
             Obtain batch_size without using run_eagerly option in compile(). This will allow the step to be run in graph mode, thereby speeding up computation.
         """
-        # Set in the UBPCallbacks() callback.
         # Set in the UBPCallbacks() callback.
         y = self._y
 
