@@ -13,10 +13,11 @@ from utils.misc import get_processor_name
 from utils.misc import generate_012_genotypes
 
 # Custom module imports
-from read_input.read_input import GenotypeData
+from snpio import GenotypeData
 from read_input.simgenodata import SimGenotypeData
 from impute.estimators import *
 from impute.simple_imputers import *
+
 
 def main():
     """[Class instantiations and main package body]"""
@@ -63,7 +64,6 @@ def main():
 
     if args.phylip:
         if args.pop_ids or args.onerow_perind:
-
             print(
                 "\nPhylip file was used with structure arguments; ignoring "
                 "structure file arguments\n"
@@ -78,38 +78,31 @@ def main():
             popmapfile=args.popmap,
             guidetree=args.treefile,
             qmatrix_iqtree=args.iqtree,
-            siterates_iqtree=args.rates
+            siterates_iqtree=args.rates,
         )
 
-        prefix="c0.001_s0.009_gtrgamma_i0.0"
-        sim = SimGenotypeData(data,
-                prop_missing=0.1,
-                strategy="random")
+        prefix = "c0.001_s0.009_gtrgamma_i0.0"
+        sim = SimGenotypeData(data, prop_missing=0.1, strategy="random")
 
         nmf = ImputeNMF(genotype_data=sim)
 
         accuracy = sim.accuracy(nmf)
-        print("Accuracy:",accuracy)
+        print("Accuracy:", accuracy)
 
         phylo = ImputePhylo(genotype_data=sim, save_plots=False)
 
         accuracy = sim.accuracy(phylo)
-        print("Accuracy:",accuracy)
+        print("Accuracy:", accuracy)
 
         nlpca = ImputeNLPCA(
-            genotype_data=sim,
-            initial_strategy="populations",
-            cv=5
+            genotype_data=sim, initial_strategy="populations", cv=5
         )
         accuracy = sim.accuracy(nlpca)
-        print("Accuracy:",accuracy)
+        print("Accuracy:", accuracy)
 
-        ubp = ImputeUBP(
-            genotype_data=sim,
-            initial_strategy="populations"
-        )
+        ubp = ImputeUBP(genotype_data=sim, initial_strategy="populations")
         accuracy = sim.accuracy(ubp)
-        print("Accuracy:",accuracy)
+        print("Accuracy:", accuracy)
 
         # vae = ImputeVAE(
         #     genotype_data=sim,
@@ -117,8 +110,6 @@ def main():
         # )
         # accuracy = sim.accuracy(vae)
         # print("Accuracy:",accuracy)
-
-
 
 
 def get_arguments():
